+++ CSC202BankSim/csc202RutiglianoLaffoonStangelBankSim/ArrivalEvent.java
@@ -1,4 +1,4 @@
-package csc202banksim;
+package csc202RutiglianoLaffoonStangelBankSim;
 
 public class ArrivalEvent extends Event{
 	private int waitTime;
