+++ CSC202BankSim/csc202RutiglianoLaffoonStangelBankSim/DepartureEvent.java
@@ -1,4 +1,4 @@
-package csc202banksim;
+package csc202RutiglianoLaffoonStangelBankSim;
 
 public class DepartureEvent extends Event {
 	private int tellerNum;
