+++ CSC202BankSim/csc202RutiglianoLaffoonStangelBankSim/RandomTime.java
@@ -1,4 +1,4 @@
-package csc202banksim;
+package csc202RutiglianoLaffoonStangelBankSim;
 
 public class RandomTime {
