+++ CSC202BankSim/csc202RutiglianoLaffoonStangelBankSim/Teller.java
@@ -1,4 +1,4 @@
-package csc202banksim;
+package csc202RutiglianoLaffoonStangelBankSim;
 
 public class Teller {
